ebuild-writing/common-mistakes: Mention the 'doc' USE flag
When reminding folks to install the docs, also remind them that they
should be using the 'doc' USE flag for serious docs.  This USE flag is
the standard way to enable/disable doc installation, and we don't want
ebuilds blindly pulling in doc depenencies regardless of how it's set.
Use some of Ulrich Müller's wording [1].

[1]: https://bugs.gentoo.org/show_bug.cgi?id=522800#c3

Thanks-to: Ulrich Müller <ulm@gentoo.org>