Makefile: construct the full dependency tree instead of pattern matching
Currently, the Makefile assumes that each document can be
independently transformed and thus exposes a flat dependency
hierarchy. This is incorrect because of the way table of contents
(TOC) is generated. When a page is added or removed, its immediate
parent and all of the parent's parents, all the way up to the root
node, need to be recompiled in a bottom-up fashion to regenerate the
TOC in each page.

Use black magic to automatically generate build rules with proper
prerequisites during runtime. The idea is to mirror the hierarchy in the
document: every section has a dependency to its subsections. The rules
are generated by iterating through each directory in the entire file
system tree.

Signed-off-by: Göktürk Yüksek <gokturk@binghamton.edu>