appendices/common-problems: remove the misleading sandbox violation example
In the "Handling Access Violations" section, there is an example of a
build system causing a sandbox violation by attempting to access
HOME. Per PMS, HOME is a predefined read-only environment variable
that points to a temporary location, which doesn't cause any sandbox
violations. Remove the example while keeping the idea of tricking the
build system into using a safer location.

Signed-off-by: Göktürk Yüksek <gokturk@binghamton.edu>