tools-reference/find: encourage the while read loop with '-print0'
The "funky while read loop" has problems coping with file names
containing special characters. Improve it by using the '-print0'
option that uses the NUL character to delimit the output lines, which
can be processed safely by the while read loop by adjusting the
delimiter settings.

As for whether supplying $'\0' is an undocumented feature or not, see:
http://lists.gnu.org/archive/html/help-bash/2016-10/msg00039.html

Suggested-By: David Seifert <soap@gentoo.org>