ebuild-writing/misc-files/metadata: repoman does not depend on libxml2 anymore
When repoman was a part of sys-apps/portage, it didn't pull in libxml2
in the ebuild. The developers needed to make sure that libxml2 is
installed on the system to perform metadata checks. Now that it's
split into app-portage/repoman, it pulls in dev-python/lxml properly
and this statement is no longer valid.