ebuild-maintenance: Remove unagreed mentions of Gentoo-Bug
Given that there is no consensus to use Gentoo-Bug in the commit
messages, I find it entirely inappropriate for the devmanual maintainers
to unilaterally push their preferred variant forward. Instead, remove
all the notions and explain the fact that there is no agreement.

This change clears the ground for future patches. In particular, it is
meant to provide the opportunity to debate a proper change
as a clear-cut solution, rather than giving the false suggestion that
it is changing the status quo.

While at it, remove the mention of Package-Manager and RepoMan-Options
footers. Users are not supposed to use those footer tags directly,
and their contents are implementation-defined.