ebuild-writing/variables: weaken the warning against third-party mirrors.
The warning against using third-party mirrors used to say that "the
mere existence of a third-party mirror list is not a good enough
reason to use it." That goes against a few existing uses of the
feature in the ::gentoo repository, so the wording is too strong. This
revision changes the warning, leaving the decision up to the
maintainer after alerting him to the potential maintenance burden.

Closes: https://bugs.gentoo.org/635646
Suggested-by: Michael Palimaka <kensington@gentoo.org>
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Signed-off-by: Brian Evans <grknight@gentoo.org>