Makefile: remove BASH-specific "type -p" and "&>" idioms.
The Makefile uses two "type -p" commands to determine if the "convert"
and "xsltproc" commands are present. The "type" command itself is
defined in POSIX,

  http://pubs.opengroup.org/onlinepubs/9699919799/utilities/type.html

but the additional "-p" flag is BASH-specific. This can lead to
unexpected behavior when the /bin/sh symlink that the Makefile uses by
default points to a non-BASH shell:

  $ dash
  $ type -p convert
  -p: not found
  convert is /usr/bin/convert

By chance, this is ultimately not fatal, but does cause the default
target to output some confusing messages. And in fact the output from
the "type -p" command is never used, which means that "type" itself
should suffice, in any shell. Thus this commit drops the two "-p"
arguments to "type".

The same two "type" commands attempt to redirect both stdout and
stderr to /dev/null using the BASH "&>" shortcut. This commit replaces
it with the standard, but more verbose incantation ">/dev/null 2>&1"
that is portable to other shells.

Closes: https://bugs.gentoo.org/680932
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Signed-off-by: Brian Evans <grknight@gentoo.org>