function-reference/install-functions: clarify dodir/keepdir usage.
The existing "Install Functions Reference" documentation on dodir()
and keepdir() is a bit wishy-washy. What developers really want to
know is, when do I use one, and when do I use the other?

This commit updates the descriptions of those two functions to provide
that information. If the directory will be non-empty, dodir() is fine.
Otherwise, keepdir() should be used. Some rationale for the existence
of the keepdir() function is also provided for the skeptical developer
who might otherwise find it superfluous.

This change mirrors an update to the ebuild(5) man page in portage.

Closes: https://bugs.gentoo.org/451472
Reported-by: Matthew Schultz <mattsch@gmail.com>
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>