general-concepts/mirrors: rework the bad upstream distfile renaming example
If a distfile rename is necessary to cope with an upstream updating a
versioned distfile in place, suggest the format of '_YYYYMMDD' instead
of '.r*'. The rationale is that the '.r*' format may confuse readers
with the Gentoo ebuild revision format '-r*', for which there is no
necessary correlation. The timestamp format also has the added bonus
of quickly being able to tell when the distfile rename took place,
assuming the developer used the proper date for it.

Signed-off-by: Göktürk Yüksek <gokturk@gentoo.org>