general-concepts/mirrors: improve bad upstream distfile rename text
If an upstream updates a distfile in a way that requires an ebuild
revision bump, remind readers to remove the ebuilds associated with
the old distfile. This is necessary because the Manifest entry for the
old tarball has its checksum changed due to upstream and cause
checksum mismatches during merges.

Signed-off-by: Göktürk Yüksek <gokturk@gentoo.org>