ebuild-writing/file-format: discourage _pre for snapshots
Do not discourage people to use the format _preYYYYMMDD when creating
snapshot ebuilds. This requires the developer to guess the version of
the next release by the upstream. The same snapshot can be treated as
a patch release over an existing version with the _pYYYYMMDD format,
which is more reliable.

Signed-off-by: Göktürk Yüksek <gokturk@gentoo.org>