Makefile: use rsvg-convert to generate PNGs from SVGs.
Our Makefile currently uses inkscape (via imagemagick's "convert"
wrapper) to turn SVGs into PNGs. This has a few downsides:

  * media-gfx/inkscape is a huge dependency,
  * inkscape is slow to perform the conversion,
  * installing imagemagick[svg] doesn't ensure that "convert" can
    convert an SVG to a PNG if inkscape is missing (bug 684686).

The good news is that there is another SVG -> PNG conversion tool
called rsvg-convert that is installed by gnome-base/librsvg. Compared
to inkscape, librsvg is a light dependency and rsvg-convert is
lightning-fast. Moreover, installing gnome-base/librsvg guarantees
that the latter will exist and can actually perform the conversion.

Inkscape is supposedly more powerful, but so far in the devmanual
we're only using SVG to put words in boxes. The simpler tool is
sufficient for that, and therefore preferable.

This commit updates the Makefile to use rsvg-convert.

Bug: https://bugs.gentoo.org/684686
Closes: https://bugs.gentoo.org/700904
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Signed-off-by: Ulrich Müller <ulm@gentoo.org>