search.js: escape HTML/XML tags returned in search results
build_search_documents.py unescapes the escaped tags when creating an
index. This is desired as lunr doesn't index them otherwise. For
example it indexes '<warning>' properly but not
'&lt;warning&gt;'. When we display them on the browser though, we need
to escape them again so that they are not interpreted as real tags.

Signed-off-by: Göktürk Yüksek <gokturk@gentoo.org>