ebuild-writing/users-and-groups: use reals examples from the tree.
The number "123" is used as the UID/GID in our example acct-user and
acct-group ebuilds. Those IDs are now however claimed by the "ntp"
user and group, so they are perhaps not the best examples. Instead,
let's just use a good example from the tree. The acct-user/group
ebuilds for suricata closely match the examples already given, but
with the added benefit of being real examples.

Since we (and the ebuilds themselves) state that they're for suricata,
it should be clear to the reader that the variables need to be
updated.  And the fact that the IDs dwell in the high-4xx range
reflects the fact that we're generally counting down from there.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Closes: https://bugs.gentoo.org/702508
Signed-off-by: Ulrich Müller <ulm@gentoo.org>