general-concepts/dependencies: Improve PDEPEND
Rewrite the PDEPEND description to be more focused on its purpose.
Rename to 'Post Dependencies' to match PMS. Reverse the weird backwards
logic that indicated that they might be installed before the package
into saying that they may be installed after it. Clearly indicate
that its only purpose is to resolve circular dependencies.

Signed-off-by: Michał Górny <mgorny@gentoo.org>