bin/gen-eclass-html.sh: More man2html postprocessing.
Remove all links containing "${", because most likely these are
instances of SRC_URI containing a variable reference.

Remove <TT> since it is no longer valid in HTML5. man2html only
outputs this pair of tags (with a sequence of non-breaking spaces
inside) when it encounters a tab character in its input. Neither
<code> nor any other logical markup would be a good replacement,
so simply remove it.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>