devmanual.css, devbook.xsl: Drop td.devbook.
Use the bootstrap default instead. Custom handling for <th> has
already been removed in commit a429f4c, drop it for <ti> as well.

About the history of this:
The vertical alignment was originally added to resolve bug 53500,
see the variables table (e.g. the KEYWORDS line) in the archived page:
https://web.archive.org/web/20040604000628/http://www.gentoo.org/doc/en/gentoo-howto.xml
The class was later (commit cff18db) renamed from tableinfo to devbook.
None of this seems to be necessary any more; vertical alignment is
just fine with the bootstrap layout.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>