*.svg: specify Open Sans instead of generic "sans-serif" font.
The text in most of our diagrams was left to use the generic
"sans-serif" font face, which depends on both the local machine and
the program used to edit/convert the SVG image. This isn't the best
situation, because we want the output PNGs to be predictable.

The Tyrian layout used for devmanual.gentoo.org makes heavy use of the
"Open Sans" font, so this commit updates each SVG in the devmanual to
use that face for all text elements that were previously "sans-serif".

Bug: https://bugs.gentoo.org/700904
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
[Fonts in ebuild-writing/functions/diagram.svg fixed]
Signed-off-by: Ulrich Müller <ulm@gentoo.org>