Makefile: New target "delete-old".
This will delete any orphaned html and image files that could be
left over after the corresponding source file was moved or removed.
Such files are also not covered by "clean", so add the new target as
prerequisite.

Run a single "find" pass to get a list of all files in ALL_FILES and
subsequently extract the necessary subsets from it. The list of all
files is needed anyway for deletion of orphans, which are not included
in HTMLS or IMAGES.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>