Makefile, depend.xsl: Use XSLT to generate the list of dependencies.
Each HTML file must depend on its XML file with all its descendants
(for the contents tree), all its ancestors (for breadcrumbs), and the
previous and next documents (for backward and forward links).

Trying to generate the list of dependencies with make seems hopeless
(especially, how would one determine the previous and next documents?).
OTOH, the necessary templates already exist in devbook.xsl. Therefore,
add a simple depend.xsl on top of it, which outputs a dependency list
as plain text.

Closes: https://bugs.gentoo.org/710918
Signed-off-by: Ulrich Müller <ulm@gentoo.org>