keywording: Use "Pure Python" instead of "Pure Perl"
We prefer not to use ALLARCHES for Perl packages because
of how sensitive Perl can be to the underlying host.

While Perl is cross-platform, use of functions which
vary across hosts (e.g. unpack) is quite common.

Let's use Python as an example instead which is
endorsed [0] by Gentoo's Python Project.

[0] https://wiki.gentoo.org/wiki/Project:Python#ALLARCHES
Signed-off-by: Sam James <sam@gentoo.org>