eclass-writing: mention that checking (official) overlays may be useful
This is just a politeness/interesting thought which people may want to consider,
rather than a hard requirement.

For "official" overlays in particular, a Gentoo developer should be able to give
the responsible developer(s) a quick ping to ensure they're aware if their overlay
seems to be heavily using legacy functionality.

Signed-off-by: Sam James <sam@gentoo.org>