general-concepts/dependencies: improve explaination of slot operators
The section previously had hard to parse sentences like "[…] and that
your package should be rebuilt if the version best matching the
runtime dependency is updated to a version with a different slot or
subslot" for the := operator.

This commit attempts to improve the readability of the section.

The change superseeds my proposed change in bug #739858.

Thanks to Sam James for reviewing this. :)

Signed-off-by: Florian Schmaus <flow@gentoo.org>
Bug: https://bugs.gentoo.org/739858
Closes: https://github.com/gentoo/devmanual/pull/259
Signed-off-by: Sam James <sam@gentoo.org>