keywording: mention filing bugs for package.mask entries
It's useful to have a bug filed for things masked for testing / due
to breakages so that feedback can be gathered in one place / the relevant
issue can be debugged.

We've sometimes had things p.masked indefinitely for "testing" when
it's not clear exactly what needs to be done left, or due to a "bug"
which with few details cannot be reproduced some time later.

Bug: https://bugs.gentoo.org/705394
Signed-off-by: Sam James <sam@gentoo.org>