ebuild-writing/misc-files/metadata: Update restrict attribute
The description currently says:
"When this attribute is used, other tags with or without the restrict
attribute must be present to cover all the versions of the package.
A tag without the restrict attribute serves as the default."

However, this is not reflected by the specification in GLEP 68.
It appears to have originated from the old DevRel Handbook where it
had been added in 2004:
https://gitweb.gentoo.org/archive/proj/gentoo.git/tree/xml/htdocs/proj/en/devrel/handbook/hb-guide-metadata.xml?id=26d18f5f20f00e538b7af0d5a795008f3426d243#n135

Signed-off-by: Ulrich Müller <ulm@gentoo.org>