general-concepts/ebuild-revisions: clarify interaction with adding/removing USE flags
Note that when adding a USE flag where the functionality was previously on
(not off in build system), one must revbump
because a dep was unconditional that now became conditional.

Also, if removing a USE flag where the behaviour is now on, but was previously
always off (not on in build system), one must revbump also,
because people might have never set USE=foo (hence --changed-use wouldn't pick
it up) but their deps have now effectively changed.

Bug: https://github.com/pkgcore/pkgcheck/pull/435
Signed-off-by: Sam James <sam@gentoo.org>