devbook.xsl: Fix links created by contentsTree extraction
If the element that is extracted from the external document contains
internal references, these links would be relative to the wrong start
location, namely the one defined in that document. Therefore, define
our own copy of /guide/@self (pointing to the location of the target
document), and rebase the tree onto it.

This fixes the intra-document links in appendices/todo-list/.

Closes: https://bugs.gentoo.org/916523
Signed-off-by: Ulrich Müller <ulm@gentoo.org>