g-c/dependencies: Clarify the stance on circular dependencies
I had a package that needed changes that would introduce a dependency
cycle with USE=test and I had to ask on IRC whether this was OK. Sam
told me not to worry about it and recommended that the devmanual
should note this.

So, add a paragraph that discourages dependency cycles but expresses
that cycles provoked by USE=test can usually be ignored.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>