general-concepts/ebuild-revisions: mention balancing examples against stable
Explicitly say that even 'no revision bump' examples should be balanced
against whether a stable ebuild is being modified, as we want to minimize
any risk of regressions there.

i.e. It might be that a change would not need a revbump for an ~arch-only
ebuild, but if the ebuild is stable, the principle of being conservative
ofr stable would mean a revbump is warranted.

Signed-off-by: Sam James <sam@gentoo.org>
Signed-off-by: Ulrich Müller <ulm@gentoo.org>