devbook-guide: Clarify intra-document references
Apparently, there is confusion about the syntax of intra-document
references; sometimes contributors copy the fragment identifier from
HTML output.

For now, this would produce correct links in HTML output, but it
relies on the properties of the conversion algorithm presently used
in the XSL stylesheet. However, the algorithm was adjusted at least
twice in the past (commit 1fb5bbc, commit 200818b) and may be changed
again in the future.

Therefore, clarify that the "#" separator must be followed by the
literal section title.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>